Rename Normalize and move to nn
