+++ examples/stable_diffusion.py
@@ -5,7 +5,6 @@ import os
 import gzip
 import argparse
 import math
-import os
 import re
 from functools import lru_cache
 from collections import namedtuple
@@ -14,36 +13,14 @@ import numpy as np
 from tqdm import tqdm
 
 from extra.utils import fake_torch_load_zipped, get_child
-from tinygrad.nn import Conv2d, Linear
+from tinygrad.nn import Conv2d, Linear, GroupNorm
 from tinygrad.tensor import Tensor
 
 # TODO: refactor AttnBlock, CrossAttention, CLIPAttention to share code
 
-# TODO: rename to GroupNorm and put in nn.py
-class Normalize:
-  def __init__(self, in_channels, num_groups=32):
-    self.weight = Tensor.empty(in_channels)
-    self.bias = Tensor.empty(in_channels)
-    self.num_groups = num_groups
-
-  def __call__(self, x):
-    # reshape for layernorm to work as group norm
-    # subtract mean and divide stddev
-    if self.num_groups == None:  # just layernorm
-      x = x.layernorm()
-    else:
-      x = x.reshape(x.shape[0], self.num_groups, -1).layernorm().reshape(x.shape)
-
-    # elementwise_affine on channels
-    if len(x.shape) == 4:
-      # HACK for channels in conv
-      return (x * self.weight.reshape(1, -1, 1, 1)) + self.bias.reshape(1, -1, 1, 1)
-    else:
-      return x.linear(self.weight, self.bias)
-
 class AttnBlock:
   def __init__(self, in_channels):
-    self.norm = Normalize(in_channels)
+    self.norm = GroupNorm(in_channels)
     self.q = Conv2d(in_channels, in_channels, 1)
     self.k = Conv2d(in_channels, in_channels, 1)
     self.v = Conv2d(in_channels, in_channels, 1)
@@ -73,9 +50,9 @@ class AttnBlock:
 
 class ResnetBlock:
   def __init__(self, in_channels, out_channels=None):
-    self.norm1 = Normalize(in_channels)
+    self.norm1 = GroupNorm(in_channels)
     self.conv1 = Conv2d(in_channels, out_channels, 3, padding=1)
-    self.norm2 = Normalize(out_channels)
+    self.norm2 = GroupNorm(out_channels)
     self.conv2 = Conv2d(out_channels, out_channels, 3, padding=1)
     self.nin_shortcut = Conv2d(in_channels, out_channels, 1) if in_channels != out_channels else lambda x: x
 
@@ -108,7 +85,7 @@ class Decoder:
       if i != 0: arr[-1]['upsample'] = {"conv": Conv2d(s[0], s[0], 3, padding=1)}
     self.up = arr
 
-    self.norm_out = Normalize(128)
+    self.norm_out = GroupNorm(128)
     self.conv_out = Conv2d(128, 3, 3, padding=1)
 
   def __call__(self, x):
@@ -142,7 +119,7 @@ class Encoder:
     self.down = arr
 
     self.mid = Mid(512)
-    self.norm_out = Normalize(512)
+    self.norm_out = GroupNorm(512)
     self.conv_out = Conv2d(512, 8, 3, padding=1)
 
   def __call__(self, x):
@@ -175,7 +152,7 @@ class AutoencoderKL:
 class ResBlock:
   def __init__(self, channels, emb_channels, out_channels):
     self.in_layers = [
-      Normalize(channels),
+      GroupNorm(channels),
       Tensor.silu,
       Conv2d(channels, out_channels, 3, padding=1)
     ]
@@ -184,7 +161,7 @@ class ResBlock:
       Linear(emb_channels, out_channels)
     ]
     self.out_layers = [
-      Normalize(out_channels),
+      GroupNorm(out_channels),
       Tensor.silu,
       lambda x: x,
       Conv2d(out_channels, out_channels, 3, padding=1)
@@ -248,9 +225,9 @@ class BasicTransformerBlock:
     self.attn1 = CrossAttention(dim, dim, n_heads, d_head)
     self.ff = FeedForward(dim)
     self.attn2 = CrossAttention(dim, context_dim, n_heads, d_head)
-    self.norm1 = Normalize(dim, num_groups=None)
-    self.norm2 = Normalize(dim, num_groups=None)
-    self.norm3 = Normalize(dim, num_groups=None)
+    self.norm1 = GroupNorm(dim, num_groups=None)
+    self.norm2 = GroupNorm(dim, num_groups=None)
+    self.norm3 = GroupNorm(dim, num_groups=None)
 
   def __call__(self, x, context=None):
     x = self.attn1(self.norm1(x)) + x
@@ -260,7 +237,7 @@ class BasicTransformerBlock:
 
 class SpatialTransformer:
   def __init__(self, channels, context_dim, n_heads, d_head):
-    self.norm = Normalize(channels)
+    self.norm = GroupNorm(channels)
     assert channels == n_heads * d_head
     self.proj_in = Conv2d(channels, n_heads * d_head, 1)
     self.transformer_blocks = [BasicTransformerBlock(channels, context_dim, n_heads, d_head)]
@@ -343,7 +320,7 @@ class UNetModel:
       [ResBlock(640, 1280, 320), SpatialTransformer(320, 768, 8, 40)],
     ]
     self.out = [
-      Normalize(320),
+      GroupNorm(320),
       Tensor.silu,
       Conv2d(320, 4, kernel_size=3, padding=1)
     ]
@@ -433,9 +410,9 @@ class CLIPAttention:
 class CLIPEncoderLayer:
   def __init__(self):
     self.self_attn = CLIPAttention()
-    self.layer_norm1 = Normalize(768, num_groups=None)
+    self.layer_norm1 = GroupNorm(768, num_groups=None)
     self.mlp = CLIPMLP()
-    self.layer_norm2 = Normalize(768, num_groups=None)
+    self.layer_norm2 = GroupNorm(768, num_groups=None)
 
   def __call__(self, hidden_states, causal_attention_mask):
     residual = hidden_states
@@ -479,7 +456,7 @@ class CLIPTextTransformer:
   def __init__(self):
     self.embeddings = CLIPTextEmbeddings()
     self.encoder = CLIPEncoder()
-    self.final_layer_norm = Normalize(768, num_groups=None)
+    self.final_layer_norm = GroupNorm(768, num_groups=None)
 
   def __call__(self, input_ids):
     x = self.embeddings(input_ids, list(range(len(input_ids))))
